Tabs: ensure to have folder icons for remote folders
KIO::iconForUrl relies on filename to find the icon for remote files.
Appending / at the end of the filename allows it to interpret it as
a folder.

This causes inconsistent tab icons.

To reproduce:

Open in a tab a url on a smb or sftp folder, i.e sftp:/my-server/photos

Before:
Icon is a cloud.

After:
Icon is a folder

Previously a folder icon could be presented if the tab was first opened
with a url with a trailing /