Fix PlacesItemModelTest, 2nd try
This partly reverts 15a5189b1193cc7305ad2c556d67ec79bdb6e38a

The KDE CI does not have KDE_FULL_SESSION set, which means that KIO
doesn't add the recentlyused:// urls to the model, but defaults to the
old timeline:// urls instead.

In both cases KIO adds 2 places, so it is safe to always assume that the
expected model count is 14.