Increase default window size
The current default window size of 750x500 is quite small on today's
screens, and results in an initial user experience of a very cramped app
window. The very first action that I see from new users interacting with
Dolphin is to make the window bigger.

So let's do it for them by making the default window size a bit bigger
ourselves.