fix padding in places view
padding was only applied when the icon size was applied, the icon size
however is only applied when the user had set an explicit size. this
resulted in inconsistent spacing. by default no padding would be used if
the user had changed the icon size to medium and back to small it would
suddenly have padding.

to fix this, set padding unconditionally on construction and never touch
it again .

BUG: 435731