Remove the filter action from the toolbar
This commit removes the "toggle_filter" action from the default
toolbar configuration. It is not quite important enough to be
always visible on the toolbar by default. Instead it is easily
accessible through the hamburger menu and the menu bar.

The filter is only really necessary if one has a lot of items in
the current location and one knows exactly how to identify a
certain item among them by the name which probably isn't how most
people browse their files.

There are also some more advanced use-cases but advanced users
will be able to either learn the keyboard shortcut or move this
action back to the toolbar.