Remove new lines when renaming a file
It is possible to paste a new filename containing new lines when renaming a
file. The new lines are kept within the new filename:
$ ls
'new'$'\n''name'

It is reasonnable to assume that, when using Dolphin to rename files,
it is not the intended behavior; removing the new lines seems like the
right thing to do.

BUG: 444747