Display Apply button in KCMs
When displaying the KCMs in another application's configuration dialog
(for example, in Konqueror), the Apply, Ok and Reset buttons aren't
shown. This issue is caused by a missing Apply flag in the call to
setButtons in the KCModule constructor.

the fact that in the call to setButtons
from the KCModule constructor, the Apply flag is missing.

BUG: 448352