Always select items on activation
There is an unintended side-effect in
d3839617193e92463806580699caa595c892b8a6 which this MR fixes.

Normally in Dolphin, when clicking on an item to open/activate it,
it is both selected by the click and opened/activated.

Prior to this MR, the item wasn't selected when all of these
conditions were met:
- Use ActivateItemOnSingleClick also knwon as single-click mode
- Have more than one item selected already
- Click on an item that was previously not selected

Prior to this MR, the click would deselect all items and
activate the clicked item but not select it.

With this MR, the click will deselect all items, activate
the clicked item and also select it and move the anchor there.

When testing this with folders, make sure to navigate back after activating the folder. The folder should then still be selected.

This is then consistent with the behaviour when the specific
conditions mentioned above are not met.