[ToolTipManager] Create DolphinFileMetaDataWidget on-demand again
This reverts b7fbd19a7660424e6839df37f12b33111408b1fc but creates the
widget only once and then reuses it.

However, ownership of a parent-less widget is transferred to the
`KToolTipWidget` but since we first request metadata, then show the
tooltip (or won't if the mouse moved on since), we cannot rely on
the `KToolTipWidget` exclusively.

Instead, when we still delete the widget ourself until we have
shown the `KToolTipWidget` once at which point it will handle the
life time for us.

This fixes parenting the widget to `DolphinView` which would make
it appear as a broken line in the top left of the window.