Improve details mode ctrl-press rubberband creation
Since d3839617193e92463806580699caa595c892b8a6 dragging the
highlighted row of already selected items in the details view mode
will begin a drag operation of all selected items.

As a unintended side-effect of this change, dragging the row of a
previously unselected item while holding the control key would also
begin a drag operation.

After this commit, dragging the row of a previously unselected item
while holding the control key in details view mode will instead
create a rubberband. Ctrl-dragging the item's icon or text directly will drag the item as expected.

With this change, using multiple rubberbands to select scattered
items among a list of items should be as convenient as it was
previously.

BUG: 452181