fix: service-menu-installer ignoring symlinks on un-install
service-menu-installer ignores <package-name>.tar.gz-dir/uninstall.sh
if this is actualy just a symbolic link to install.sh in the same directory
(hardlinks and copies work)

canonicalPath() is used to get the name of the file on disk and returns install.sh
in the case of a symbolic link. install.sh without any arguments is called instead
of uninstall.sh consequently and no de-installation takes place.

replace canonicalPath() with absolutePath() as used in the rest of source file

resolves:

BUG: 452289