Re-add "Open Terminal Here" feature
This is equivalent to the "Open Terminal Here" feature that existed until Version 20.12.  

If the user has selected folders, replace "Open Terminal" in the context menu with "Open Terminal Here".  
When more than 5 folders are selected, a modal window will ask the user if they are sure they want to
open all 6 or more terminal windows.

In Detail View, users can also select a file, 
which will open a terminal at the location of that file.

BUG: 452637
FIXED-IN: 22.08