Fix paste on row while in details view mode
Before this change, right-clicking the row of an unselected item
in details view mode would be in a weird state:
- It didn't really count as a click on the item because the item
    didn't get selected by this click before opening the context
    menu.
- It didn't really count as a click on the view background either
    because the actions that showed up depended on the item in
    that row.

This commit fixes this by considering a right-click in the same row
as an unselected item as a click on the view background.
The behaviour of right-clicking the icon or name of a file directly
is unchanged.

This fixes the following bugs:
- The Paste action that shows up when right-clicking in the
    unselected row of a folder now works (instead of doing
    nothing). It now pastes the clipboard contents onto the view
    background.
- When right-clicking the unselected row of a file (not a folder)
    a Paste action once again shows up.