Remove hover highlight before opening view context menu
If one was fast to open the right-click context menu on the row of
an item in details view mode, the hover highlight would persist
while the context menu for the view was open.

This one-liner makes it so the highlight on the row is always
removed before the right-click context menu for the view is opened
so it is as clear as possible that the newly opened context menu
has no relation to the fileItem.

(cherry picked from commit f7365aa9ca9d1242cb2a0c4505c64e2ab05e0d67)