Extend selection rectangle to left edge
This commit changes it so the sizes of selection rectangles and hover
highlights in compact and details view mode is identical for all items.

Before this commit, selection rectangles in lists would have varying
indentation of the left edge of the selection rectangle depending on
the preview image's width-to-height ratio. This would cause a sort of
"ragged edge" in both compact and details list view when multiple items
were selected.

This commit doesn't change anything about icon view mode.

BUG: 453046